Move account defaults to AccountDefaultsProvider
AccountPreferenceSerializer is not a good place to host these
