+++ app-common/build.gradle.kts
@@ -9,6 +9,7 @@ android {
 dependencies {
     api(projects.legacy.common)
 
+    implementation(projects.legacy.core)
     implementation(projects.legacy.account)
 
     implementation(projects.core.featureflags)
