+++ app-common/src/main/kotlin/net/thunderbird/app/common/AppCommonModule.kt
@@ -6,5 +6,10 @@ import org.koin.core.module.Module
 import org.koin.dsl.module
 
 val appCommonModule: Module = module {
-    single<AccountDefaultsProvider> { CommonAccountDefaultsProvider(featureFlagProvider = get()) }
+    single<AccountDefaultsProvider> {
+        CommonAccountDefaultsProvider(
+            resourceProvider = get(),
+            featureFlagProvider = get(),
+        )
+    }
 }
