+++ legacy/core/src/main/java/com/fsck/k9/AccountPreferenceSerializer.kt
@@ -1,8 +1,6 @@
 package com.fsck.k9
 
 import app.k9mail.legacy.account.Account
-import app.k9mail.legacy.account.Account.Companion.DEFAULT_SORT_ASCENDING
-import app.k9mail.legacy.account.Account.Companion.DEFAULT_SORT_TYPE
 import app.k9mail.legacy.account.Account.Companion.DEFAULT_SYNC_INTERVAL
 import app.k9mail.legacy.account.Account.Companion.NO_OPENPGP_KEY
 import app.k9mail.legacy.account.Account.Companion.UNASSIGNED_ACCOUNT_NUMBER
@@ -25,7 +23,6 @@ import com.fsck.k9.preferences.StorageEditor
 import timber.log.Timber
 
 class AccountPreferenceSerializer(
-    private val resourceProvider: CoreResourceProvider,
     private val serverSettingsSerializer: ServerSettingsSerializer,
 ) {
 
@@ -556,81 +553,6 @@ class AccountPreferenceSerializer(
 
                 defaultEnum
             }
-        }
-    }
-
-    fun loadDefaults(account: Account) {
-        with(account) {
-            automaticCheckIntervalMinutes = DEFAULT_SYNC_INTERVAL
-            idleRefreshMinutes = 24
-            displayCount = K9.DEFAULT_VISIBLE_LIMIT
-            accountNumber = UNASSIGNED_ACCOUNT_NUMBER
-            isNotifyNewMail = true
-            folderNotifyNewMailMode = FolderMode.ALL
-            isNotifySync = false
-            isNotifySelfNewMail = true
-            isNotifyContactsMailOnly = false
-            isIgnoreChatMessages = false
-            messagesNotificationChannelVersion = 0
-            folderDisplayMode = FolderMode.NOT_SECOND_CLASS
-            folderSyncMode = FolderMode.FIRST_CLASS
-            folderPushMode = FolderMode.NONE
-            sortType = DEFAULT_SORT_TYPE
-            setSortAscending(DEFAULT_SORT_TYPE, DEFAULT_SORT_ASCENDING)
-            showPictures = ShowPictures.NEVER
-            isSignatureBeforeQuotedText = false
-            expungePolicy = Expunge.EXPUNGE_IMMEDIATELY
-            importedAutoExpandFolder = null
-            legacyInboxFolder = null
-            maxPushFolders = 10
-            isSubscribedFoldersOnly = false
-            maximumPolledMessageAge = -1
-            maximumAutoDownloadMessageSize = DEFAULT_MAXIMUM_AUTO_DOWNLOAD_MESSAGE_SIZE
-            messageFormat = DEFAULT_MESSAGE_FORMAT
-            isMessageFormatAuto = DEFAULT_MESSAGE_FORMAT_AUTO
-            isMessageReadReceipt = DEFAULT_MESSAGE_READ_RECEIPT
-            quoteStyle = DEFAULT_QUOTE_STYLE
-            quotePrefix = DEFAULT_QUOTE_PREFIX
-            isDefaultQuotedTextShown = DEFAULT_QUOTED_TEXT_SHOWN
-            isReplyAfterQuote = DEFAULT_REPLY_AFTER_QUOTE
-            isStripSignature = DEFAULT_STRIP_SIGNATURE
-            isSyncRemoteDeletions = true
-            openPgpKey = NO_OPENPGP_KEY
-            isRemoteSearchFullText = false
-            remoteSearchNumResults = DEFAULT_REMOTE_SEARCH_NUM_RESULTS
-            isUploadSentMessages = true
-            isMarkMessageAsReadOnView = true
-            isMarkMessageAsReadOnDelete = true
-            isAlwaysShowCcBcc = false
-            lastSyncTime = 0L
-            lastFolderListRefreshTime = 0L
-
-            setArchiveFolderId(null, SpecialFolderSelection.AUTOMATIC)
-            setDraftsFolderId(null, SpecialFolderSelection.AUTOMATIC)
-            setSentFolderId(null, SpecialFolderSelection.AUTOMATIC)
-            setSpamFolderId(null, SpecialFolderSelection.AUTOMATIC)
-            setTrashFolderId(null, SpecialFolderSelection.AUTOMATIC)
-            setArchiveFolderId(null, SpecialFolderSelection.AUTOMATIC)
-
-            identities = ArrayList<Identity>()
-
-            val identity = Identity(
-                signatureUse = false,
-                signature = null,
-                description = resourceProvider.defaultIdentityDescription(),
-            )
-            identities.add(identity)
-
-            updateNotificationSettings {
-                NotificationSettings(
-                    isRingEnabled = true,
-                    ringtone = DEFAULT_RINGTONE_URI,
-                    light = NotificationLight.Disabled,
-                    vibration = NotificationVibration.DEFAULT,
-                )
-            }
-
-            resetChangeMarkers()
         }
     }
