+++ legacy/core/src/main/java/com/fsck/k9/notification/CoreKoinModule.kt
@@ -37,7 +37,6 @@ val coreNotificationModule = module {
     }
     single {
         AccountPreferenceSerializer(
-            resourceProvider = get(),
             serverSettingsSerializer = get(),
         )
     }
